moved handle resize/update isMobile into the appStatus provider and removed other duplicate hooks. added some comments to graph data hooks
