+++ packages/augur-comps/src/stores/app-status.tsx
@@ -1,7 +1,17 @@
-import React from 'react';
+import React, { useEffect } from 'react';
 import { DEFAULT_APP_STATUS_STATE, STUBBED_APP_STATUS_ACTIONS } from './constants';
 import { useAppStatus } from './app-status-hooks';
 import { useUserStore } from './user';
+import { windowRef } from '../utils/window-ref';
+
+function checkIsMobile(setIsMobile) {
+  const isMobile =
+    (
+      windowRef.getComputedStyle(document.body).getPropertyValue('--is-mobile') ||
+      ''
+    ).indexOf('true') !== -1;
+  setIsMobile(isMobile);
+}
 
 export const AppStatusContext = React.createContext({
   ...DEFAULT_APP_STATUS_STATE,
@@ -32,6 +42,15 @@ export const AppStatusProvider = ({ children }) => {
   const readableState = { ...state };
   delete readableState.actions;
   AppStatusStore.get = () => readableState;
+
+  useEffect(() => {
+    const handleResize = () => checkIsMobile(state.actions.setIsMobile);
+    windowRef.addEventListener('resize', handleResize);
+    handleResize();
+    return () => {
+      windowRef.removeEventListener('resize', handleResize);
+    };
+  }, []);
 
   return (
     <AppStatusContext.Provider value={state}>
