+++ packages/augur-migration/src/modules/App.tsx
@@ -25,29 +25,6 @@ const {
   AppStatus: { AppStatusProvider, useAppStatusStore },
 } = Stores;
 
-function checkIsMobile(setIsMobile) {
-  const isMobile =
-    (
-      window.getComputedStyle(document.body).getPropertyValue('--is-mobile') ||
-      ''
-    ).indexOf('true') !== -1;
-  setIsMobile(isMobile);
-}
-
-function useHandleResize() {
-  const {
-    actions: { setIsMobile },
-  } = useAppStatusStore();
-  useEffect(() => {
-    const handleResize = () => checkIsMobile(setIsMobile);
-    window.addEventListener('resize', handleResize);
-    handleResize();
-    return () => {
-      window.removeEventListener('resize', handleResize);
-    };
-  }, []);
-}
-
 const AppBody = () => {
   const { isMobile, modal } = useAppStatusStore();
   const {
@@ -59,7 +36,6 @@ const AppBody = () => {
   useUserBalances();
   useFinalizeUserTransactions();
   useUpdateApprovals();
-  useHandleResize();
   useRepMigrated();
 
   const { networkId } = PARA_CONFIG;
